lbmap: Fix padding of Service6ValueV2
The actual padding of the struct is u32 instead of u16. The mismatch
is probably harmless, as Go compiled should have added an u16 unexported
field. Otherwise, the alignchecker would have complained.

Fixes: c89f3a2725 ("lbmap: Get rid of LB weights")

+++ pkg/maps/lbmap/ipv6.go
@@ -189,7 +189,7 @@ type Service6ValueV2 struct {
 	BackendID uint32 `align:"backend_id"`
 	Count     uint16 `align:"count"`
 	RevNat    uint16 `align:"rev_nat_index"`
-	Pad       uint16
+	Pad       uint32
 }
 
 func NewService6ValueV2(count uint16, backendID loadbalancer.BackendID, revNat uint16) *Service6ValueV2 {
